Use __new__ and __init__ separately.
This sets us up for having resurrect and new be different, but doing that will
also require changing the signature of PeristentList and PersistentDict, so
take this small step first.  This also moves the guts of new into the
MemroryManager, since my experiments with PersistenObject indicate that will be
needed (PersistentObject needs to instantiate a PersistentList, and doubtless
other derived types will need to do similar things).