Handle None correctly, and add True and False.
This answers an XXX comment: it does not in the general case work to
use OID_NULL as equivalent to None.  NULL and None are not quite
interchangeable in Python, even though it mostly worked.

This changeset introduces a general way to handle singletons.  There
are a few more Python singletons to add at some point.